global bar should manage contents
